requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langgraph>=0.1.0
langsmith>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
streamlit>=1.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
arxiv>=1.4.0
duckduckgo-search>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
typing-extensions>=4.5.0
